bind-arrays {DelayedArray}	R Documentation

Bind arrays along their rows or columns

Description

Bind array-like objects with an arbitrary number of dimensions along their rows (arbind) or columns (acbind).

Usage

arbind(...)
acbind(...)

Arguments

...
The array-like objects to bind.

Value

An array-like object, typically of the same class as the input objects if they all have the same class.

See Also

DelayedArray in this package for arbind/acbind'ing DelayedArray objects.

rbind and cbind in the base package for the corresponding operations on matrix-like objects.

The abind package from CRAN.

Examples

a1 <- array(1:60, c(3, 5, 4),
            dimnames=list(NULL, paste0("M1y", 1:5), NULL))
a2 <- array(101:240, c(7, 5, 4),
            dimnames=list(paste0("M2x", 1:7), paste0("M2y", 1:5), NULL))
a3 <- array(10001:10100, c(5, 5, 4),
            dimnames=list(paste0("M3x", 1:5), NULL, paste0("M3z", 1:4)))

arbind(a1, a2, a3)

[Package DelayedArray version 0.16.3 Index]