compat {DelayedArray}	R Documentation

Functions and classes that have moved to S4Arrays

Description

Some functions and classes that used to be defined in the DelayedArray package have been moved to the new S4Arrays package in BioC 3.17. The corresponding symbols are still exported by the DelayedArray package for backward compatibility with existing code.

WARNING: This is a temporary situation only. Packages that import these symbols from DelayedArray must be modified to import them from S4Arrays instead.

These symbols are actually documented in the S4Arrays package. See:

S4Arrays::t.Array

S4Arrays::makeNindexFromArrayViewport

S4Arrays::ArrayGrid

S4Arrays::DummyArrayGrid

S4Arrays::RegularArrayGrid

S4Arrays::ArbitraryArrayGrid

S4Arrays::extract_array

S4Arrays::is_sparse

S4Arrays::read_block

S4Arrays::write_block

[Package DelayedArray version 0.36.0 Index]