chunkGrid {DelayedArray}	R Documentation

chunkGrid

Description

chunkGrid and chunkdim are internal generic functions not aimed to be used directly by the user.

Usage

chunkGrid(x)
chunkdim(x)

Arguments

x
An array-like object.

Details

Coming soon...

Value

chunkGrid returns NULL or an ArrayGrid object defining a grid on reference array x.

chunkdim returns NULL or the chunk dimensions in an integer vector parallel to dim(x).

See Also

defaultAutoGrid and family to generate automatic grids to use for block processing of array-like objects.

DelayedArray objects.

ArrayGrid for the formal representation of grids and viewports.

Examples

## Coming soon...

[Package DelayedArray version 0.16.3 Index]